extracting registerUserApi to separate file
