+++ src/services/getAllBoxes.ts
@@ -1,6 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
 import { axiosInstance } from "./apiClient";
-import { FormFields } from "src/components/login/Login";
 
 export const getAllBoxes = async () => {
   const response = await axiosInstance.get("/api/box/search");
@@ -13,8 +12,3 @@ export function useBoxes() {
     queryFn: getAllBoxes,
   });
 }
-
-export const registerUserApi = async (data: FormFields) => {
-  const response = axiosInstance.post("/user/register", data);
-  return response;
-};
